Completed sourcing details of all components
Sourcing data is organized in the following way:
Supplier1 - The primary supplier (also the cheapest)
Supplier1 part - The primary supplier part number
Supplier1 link - Link to the part on the suppliers homepage

The same goes for Supplier2 to 5.
All parts also have a Vendor and Vendor part field to identify
a vendor for each part.

Fixes #6